Clean-ups and some improvements

+++ src/main/java/pl/edu/pw/ee/pyskp/documentworkflow/data/domain/Difference.java
@@ -1,16 +1,15 @@
 package pl.edu.pw.ee.pyskp.documentworkflow.data.domain;
 
 import lombok.Data;
-import lombok.EqualsAndHashCode;
 
 /**
  * Created by piotr on 11.12.16.
  */
 @Data
 public class Difference {
-    private Long previousSectionStart, previousSectionSize;
+    private int previousSectionStart, previousSectionSize;
 
-    private Long newSectionStart, newSectionSize;
+    private int newSectionStart, newSectionSize;
 
     private DifferenceType differenceType;
 }
