+++ src/main/java/pl/edu/pw/ee/pyskp/documentworkflow/data/domain/FileMetadata.java
@@ -33,5 +33,5 @@ public class FileMetadata {
     @DBRef(lazy = true)
     private Version latestVersion;
 
-    private Integer numberOfVersions;
+    private long numberOfVersions;
 }
