+++ src/main/java/pl/edu/pw/ee/pyskp/documentworkflow/data/domain/Task.java
@@ -38,5 +38,5 @@ public class Task {
     @DBRef
     private FileMetadata lastModifiedFile;
 
-    private Integer numberOfFiles;
+    private long numberOfFiles;
 }
