+++ src/main/java/pl/edu/pw/ee/pyskp/documentworkflow/dtos/project/ProjectSummaryDTO.java
@@ -20,7 +20,7 @@ public class ProjectSummaryDTO {
     @NonNull
     Date creationDate;
 
-    int numberOfParticipants, numberOfTasks, numberOfFiles;
+    long numberOfParticipants, numberOfTasks, numberOfFiles;
 
     FileSummaryDTO lastModifiedFile;
