+++ src/main/java/pl/edu/pw/ee/pyskp/documentworkflow/dtos/task/TaskSummaryDTO.java
@@ -20,7 +20,7 @@ public class TaskSummaryDTO {
 
     private FileSummaryDTO lastModifiedFile;
 
-    private int numberOfFiles, numberOfParticipants;
+    private long numberOfFiles, numberOfParticipants;
 
     public static TaskSummaryDTO fromTask(Task task) {
         FileSummaryDTO lastModifiedFile = null;
